using integers for frame numbers at xml import

+++ fastspt/readers.py
@@ -66,7 +66,7 @@ def read_trackmate_xml(path):
     
     try:
         for particle in data['Tracks']['particle']:  
-            traces.append([(float(d['@x']), float(d['@y']), float(d['@t'])*framerate, float(d['@t'])) for d in particle['detection']])
+            traces.append([(float(d['@x']), float(d['@y']), float(d['@t'])*framerate, int(d['@t'])) for d in particle['detection']])
     except KeyError as e:
         print(f'problem with {path}')
         raise e
